Move HarmonyExceptionHandler to OM.Core
