+++ unturned/OpenMod.Unturned/Animals/Events/AnimalsEventsListener.cs
@@ -3,9 +3,9 @@ using System;
 using System.Reflection;
 using HarmonyLib;
 using JetBrainsAnnotations::JetBrains.Annotations;
+using OpenMod.Core.Patching;
 using OpenMod.UnityEngine.Extensions;
 using OpenMod.Unturned.Events;
-using OpenMod.Unturned.Patching;
 using SDG.Framework.Devkit;
 using SDG.Unturned;
 using Steamworks;
