+++ unturned/OpenMod.Unturned/Players/Clothing/Events/PlayerClothingEventsListener.cs
@@ -5,9 +5,9 @@ using System.Reflection;
 using HarmonyLib;
 using JetBrainsAnnotations::JetBrains.Annotations;
 using OpenMod.API.Eventing;
+using OpenMod.Core.Patching;
 using OpenMod.Unturned.Events;
 using OpenMod.Unturned.Items;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 
 namespace OpenMod.Unturned.Players.Clothing.Events
