+++ unturned/OpenMod.Unturned/Players/Equipment/Events/PlayerEquipmentEventsListener.cs
@@ -2,9 +2,9 @@
 using System.Reflection;
 using HarmonyLib;
 using JetBrains.Annotations;
+using OpenMod.Core.Patching;
 using OpenMod.Unturned.Events;
 using OpenMod.Unturned.Items;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 
 namespace OpenMod.Unturned.Players.Equipment.Events
