+++ unturned/OpenMod.Unturned/Players/Interacting/Events/UnturnedInteractEventsListener.cs
@@ -3,8 +3,8 @@ using System;
 using System.Reflection;
 using HarmonyLib;
 using JetBrainsAnnotations::JetBrains.Annotations;
+using OpenMod.Core.Patching;
 using OpenMod.Unturned.Events;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 using Steamworks;
