+++ unturned/OpenMod.Unturned/Players/Quests/Events/PlayerQuestsEventsListener.cs
@@ -1,7 +1,7 @@
 ﻿using HarmonyLib;
 using JetBrains.Annotations;
+using OpenMod.Core.Patching;
 using OpenMod.Unturned.Events;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 using Steamworks;
 using System;
