+++ unturned/OpenMod.Unturned/Vehicles/Events/VehicleEventsListener.cs
@@ -2,9 +2,9 @@
 using HarmonyLib;
 using JetBrainsAnnotations::JetBrains.Annotations;
 using OpenMod.API;
+using OpenMod.Core.Patching;
 using OpenMod.UnityEngine.Extensions;
 using OpenMod.Unturned.Events;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 using Steamworks;
 using System;
