+++ unturned/OpenMod.Unturned/Zombies/Events/ZombieEventsListener.cs
@@ -5,8 +5,8 @@ using System.Reflection;
 using HarmonyLib;
 using JetBrainsAnnotations::JetBrains.Annotations;
 using OpenMod.API;
+using OpenMod.Core.Patching;
 using OpenMod.Unturned.Events;
-using OpenMod.Unturned.Patching;
 using SDG.Unturned;
 using UnityEngine;
 // ReSharper disable InconsistentNaming
